edit: auth protection and client - user id relationship

+++ app/Http/Controllers/ClientController.php
@@ -2,17 +2,17 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Client;
 
 class ClientController extends Controller
 {
     public function index(){
-        $clients = Client::all();
+        $clients = Client::where('user_id', Auth::id())->get();
 
         return view('clients.index', [
             'clients' => $clients,
-    ]);
+        ]);
     }
 
     public function create() {
@@ -26,17 +26,28 @@ class ClientController extends Controller
             'company' => 'nullable',
             'phone' => 'nullable',
         ]);
-        Client::create($validated);
+
+        $client = new Client($validated);
+        $client->user_id = Auth::id();
+        $client->save();
+
         return redirect('/clients');
+
+        /* Client::create($validated);
+        return redirect('/clients'); */
     }
 
-    public function edit(Client $client) {
+    public function edit(Client $client)
+    {
+        abort_unless($client->user_id === Auth::id(), 403);
         return view('clients.edit', [
-            'client' => $client,
+        'client' => $client,
         ]);
     }
 
     public function update(Client $client ) {
+
+        abort_unless($client->user_id === Auth::id(), 403);
         $validated = request()->validate([
             'name' => 'required',
             'email' => 'nullable|email',
@@ -48,12 +59,16 @@ class ClientController extends Controller
     }
 
     public function destroy(Client $client) {
+
+        abort_unless($client->user_id === Auth::id(), 403);
         $client->delete();
 
         return redirect('/clients');
     }
 
     public function show(Client $client) {
+
+        abort_unless($client->user_id === Auth::id(), 403);
         $client->load('projects');
         return view('clients.show', [
             'client' => $client,
